Use the request's 'lookup' option when client-side lookup is required
Client-side lookup is used for socks4 and socks5, but not socks,
socks4a or socks5h, which delegate lookup to the remote proxy itself.

This removes the dnsCache option that was briefly added in 6.2.0-beta.0
in favour of a more generic solution using the standard Node options,
but includes a test that adds caching using this 'lookup' option as an
alternative to using dnsCache.